Add utils file for NotebookController with a logs / routine_handler class
